+++ lib/components/Dialog/DialogListItem.tsx
@@ -107,6 +107,7 @@ export const DialogListItem = memo(
     trashedAtLabel,
     badge,
     dueAt,
+    dueAtExpired,
     dueAtLabel,
     unread,
     unreadLabel = 'Ulest',
@@ -198,6 +199,7 @@ export const DialogListItem = memo(
               trashedAt={trashedAt}
               trashedAtLabel={trashedAtLabel}
               dueAt={dueAt}
+              dueAtExpired={dueAtExpired}
               dueAtLabel={dueAtLabel}
               attachmentsCount={attachmentsCount}
               attachmentsLabel={attachmentsLabel}